Defines iter_data using a factory function
