+++ SNData/csp/dr3/__init__.py
@@ -24,7 +24,6 @@ from ._data_download import download_module_data
 from ._data_parsing import get_available_ids
 from ._data_parsing import get_available_tables
 from ._data_parsing import get_data_for_id
-from ._data_parsing import get_sncosmo_input
 from ._data_parsing import iter_data
 from ._data_parsing import load_table
 from ._data_parsing import register_filters
