+++ SNData/csp/dr3/_data_parsing.py
@@ -115,27 +115,4 @@ def get_data_for_id(obj_id, format_sncosmo=False):
     return data_table
 
 
-def iter_data(verbose=False, format_sncosmo=False, filter_func=None):
-    """Iterate through all available targets and yield data tables
-
-    An optional progress bar can be formatted by passing a dictionary of tqdm
-    arguments. Outputs can be optionally filtered by passing a function
-    ``filter_func`` that accepts a data table and returns a boolean.
-
-    Args:
-        verbose  (bool, dict): Optionally display progress bar while iterating
-        format_sncosmo (bool): Format data for SNCosmo.fit_lc (Default: False)
-        filter_func    (func): An optional function to filter outputs by
-
-    Yields:
-        Astropy tables
-    """
-
-    if filter_func is None:
-        filter_func = lambda x: x
-
-    iterable = utils.build_pbar(get_available_ids(), verbose)
-    for id_val in iterable:
-        data_table = get_data_for_id(id_val, format_sncosmo=format_sncosmo)
-        if filter_func(data_table):
-            yield data_table
+iter_data = utils.create_iter_func(get_available_ids, get_data_for_id)
